observer: move entry point from lib/ to .
