+++ observer/bin/dry-run.js
@@ -2,7 +2,7 @@ import * as SparkImpactEvaluator from '@filecoin-station/spark-impact-evaluator'
 import { ethers } from 'ethers'
 
 import { RPC_URL, rpcHeaders } from '../lib/config.js'
-import { observeTransferEvents, observeScheduledRewards } from '../lib/observer.js'
+import { observeTransferEvents, observeScheduledRewards } from '../index.js'
 import { getPgPools } from '@filecoin-station/spark-stats-db'
 
 const pgPools = await getPgPools()
