+++ observer/test/observer.test.js
@@ -3,7 +3,7 @@ import { beforeEach, describe, it } from 'mocha'
 import { getPgPools } from '@filecoin-station/spark-stats-db'
 import { givenDailyParticipants } from 'spark-evaluate/test/helpers/queries.js'
 
-import { observeTransferEvents, observeScheduledRewards } from '../lib/observer.js'
+import { observeTransferEvents, observeScheduledRewards } from '../index.js'
 
 describe('observer', () => {
   let pgPools
